renaming some components
