+++ client/src/views/InternshipModule.vue
@@ -2,26 +2,26 @@
   <div class="container-fluid">
     <template v-if="loadingState !== true && internshipModule !== null">
       <!-- Kein Praktikum gefunden -->
-      <NoCompleteInternshipComponent
+      <no-complete-internship
         v-if="internshipModule.status === 'unknown'"
-      ></NoCompleteInternshipComponent>
+      ></no-complete-internship>
       <!-- Praktikum gefunden -->
-      <CompleteInternshipComponent v-else ></CompleteInternshipComponent>
+      <complete-internship v-else ></complete-internship>
     </template>
   </div>
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import NoCompleteInternshipComponent from '@/components/internship-module/NoInternshipModuleComponent.vue';
-import CompleteInternshipComponent from '@/components/internship-module/InternshipModuleComponent.vue';
+import NoCompleteInternship from '@/components/internship-module/NoInternshipModule.vue';
+import CompleteInternship from '@/components/internship-module/InternshipModule.vue';
 import http from '@/utils/http-common';
 
 export default defineComponent({
   name: 'InternshipModule',
   components: {
-    NoCompleteInternshipComponent,
-    CompleteInternshipComponent,
+    NoCompleteInternship,
+    CompleteInternship,
   },
   data() {
     return {
