+++ client/src/views/Login.vue
@@ -4,7 +4,7 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import LoginComponent from '@/components/LoginComponent.vue';
+import LoginComponent from '@/components/LoginForm.vue';
 
 export default defineComponent({
   name: 'Login',
